Refactor: Return userId type to Long

+++ backend/src/main/java/com/example/pomodoroApp/model/UserAccount.java
@@ -20,7 +20,7 @@ public class UserAccount {
     @Id
     @GeneratedValue
     @Column(updatable = false, nullable = false)
-    BigInteger userId;
+    Long userId;
 
     @Column(updatable = false, nullable = false)
     String googleUserId;
